compose nav - working

+++ ui/analysis/src/main/java/com/example/analysis/AnalysisUiState.kt
@@ -5,6 +5,6 @@ data class AnalysisUiState(
     val dailyTime: Int = 0,
     val yesterdayTime: Int = 0,
     val lastWeekAvgTime: Int = 0,
-    val lastMonthAvgTime: Int = 0,
+    val lastMonthAvgTime: Long = 0,
     val isCompleted: Boolean = false,
 )
